ship_to_address and legal_name properties have been added to Profile model.
Shipping address is needed from time to time. Anyway, there is no strict restriction whether it is specified or not. A user may want to receive all packages at their residential address. ship_to_address is a simple property that returns an actual shipping address only if it is set.

Public name is used in many places throughout the site, but sometimes it is worse than useless. Full, legal name is needed for various purposes rather than something which may be a nick name. This is why legal_name property has been added.