street_extra property added to Address model.
It will store any kind of additional information regarding address. For some people a single line is not enough. My initial plan was to add another field to profile form and store them all in the single street property, separated by a special character. It will be probably clearer, when it is split into two properties.