Some fixes to the map reduce profile conversion script.
- typo in postalcode fixed
- invalid emails are not set
(for some reason some exisiting users have invalid email addresses, like 'a' or 'foo_at_gmail_com')
- too long links are not set
(in the old DB model, LinkProperty existed. It was similar to StringProperty but there was no 500 character limit for indexed properties - or the properties were not indexed by default. This class does not exist in NDB and StringProperty should be used for links as well. String properties, which are indexed, cannot be longer than 500 characters)