Some email and phone number specific CSS is removed.
This CSS comes from old times when it was delivered by Echo Ditto. They designed a fancy form for profile registration but we ended up just iterating over the fields. It turns out that there was some special code for email and phone rows. It was applied now, thus these fields rendered in a weird field.

This change just discards the special properties and the normal rendering is restored.