Break cyclic imports among test modules.
This allows our lint coverage to expand to include
our test directories. Woohoo!

The mechanism used here is to recognize a family of
high-level seeding functions that call one another
when various keyword arguments of theirs are not
supplied by callers, and to extract from them a set
of low-level seeding functions that require all
arguments to be passed. For the moment the low-level
functions are aggregated in a "seed_utils" module
but I intend to clean up and deduplicate the code
after lint coverage is fully extended and the build
is completely passing. This is sufficient for now.