clean_birth_date function removed from the cleaning logic.
It will not be used anymore, as it is closely bound to an instance of a django form. Suffice it to look at the removed unit tests. The nice thing about this function was that it was covered well by unit tests. Anyway, it is not a big loss, because the function is actually a wrapper of isAgeSufficientForProgram function which is covered by tests even better.