ship_to_address function changed.
The downside of this change is that it constructs a new object. However, it should not be a performance issue. The good thing is that name is always set to something that makes sense. It is legal_name if no shipping address is provided or ship_to is not set.