Add pointless-statement suppressions to two test_request_data modules.
Throughout both of these modules attribute access is itself the
operation under test and the test doesn't care what the value of
the attribute is. This confuses pylint which operates with the
assumption that attribute access is safe, side-effect-free, and
generally uninteresting, and consequently any statement that is
a mere attribute access expression is a pointless statement.

In pylint's defense, its assumption is valid for all sane code
and we're the ones being weird making our attribute accesses so
complicated that they merit testing.