The tests tries to open the task page when a comment is successfully submitted.
The extended test case failed because the task view tried to access 'created_by' property directly. It is not possible right now,because it points to NDB Profile instance rather than old GCIProfile.

The backend code is now updated and fixed so the page for a task with comments loads fine.