UrlPatternConstructor interface added.
It will be used to construct Django URL patterns and eventually replace url() functions defined in these modules. The advantage of the class over raw functions is documentation especially when passed as function arguments.

There are two implementations added at this point: one for Summer Of Code and one for Code In. It will be passed to subclasses (and horepfully to the main class itself too at some point) of RequestHandler so that it will be easier to share code.

Currently, we need to implement a new subclass in order to define Django URL patterns. Each sublcass of RequestHandler is required to implement djangoURLPatterns function. At the same time, url() function must be explicitly imported from the appropriate url_patterns module. The class takes it from outside. As a result, we have to add two subclasses if the only difference is that they belong to different (GSoC or GCI) namespaces. Consider the following example:

class GSoCSpecific(base.GSoCRequestHandler):

  def djangoURLPatterns(self):
    return soc_url_patterns.url(...)

class GCISpecific(base.GCIRequestHandler):

  def djangoURLPatterns(self):
    return ci_url_patterns.url(...)

It is difficult to share other logic, even if it is the same for both programs. Multiple inheritance comes into play. It can be easily avoided if the appropriate interface to construct patterns is injected to the view. The code above shrinks to a single class:

class General(base.RequestHandler):
  def djangoURLPatterns(self):
    return self.url_pattern_constructor.constructs(...)