Page to display survey responses in read-only mode is added.
Currently, org admins are not able to see them when the org application period is over. We do not allow to access the edit view, becauce, well, we do not want them to mess with it.
This view solves the problem, because it simply lists all the answers.

A few simple unit tests are also added in this revision.