user is not created by ProfileHelper but it is injected to the object by assigning it.
It is a tricky problem. ProfileHelper class has account ID of the logged in user hardcoded. Weird, isn't it? So, even if there are a few ProfileHelper objects, their createUser() function creates User entities for the same account ID. When we started passing actual RequestData objects, user() property tried to obtain User entity based on account ID. It just took *any* of those. The tests failed.