Pass program_messages entity to the org app acceptance/rejection logic and then down to notification context providers.
In the notification context providers, instead of retrieving the
program messages within the method, we use the entity passed
on by the caller since this method gets executed within another
transaction now. We just pass on the relevant message property
from the program messages entity to the getContext method to
build the notification email object.