StartConnectionAsUser view added.
It will be used by users to connect with organizations.

Unit tests will be added at some point later. It is not possible to test this class, because it does not work yet :) Well, it almost works but a view to manage connections must be added first, because this view needs to generate a link to that page.