Send the individual message test emails transactionally instead of sending all of them together in a single enclosing transaction.
These emails are merely test emails to be sent by program admins to the
email ids of their choice. We have a problem with scheduling all the
emails together in a single enclosing transaction because because when
we get to the point when we have more than 5 messages to be emailed the
number of entity groups will exceed the number 5. App Engine has a quota
limit of being able to deal with only 5 entity groups. When we have more
than 5 messages, we will have more than 5 entity groups, so this
transaction fails. So instead of scheduling all the emails to be sent in
a single transaction we are scheduling each mail in its own transaction.

This is feature is used only by program admins for testing purposes so
as to gain enough confidence about the emails that they are about to
send. So it doesn't matter if all the emails are within a single
transaction, i.e. if part of them fail or all of them fail, they can
always retry. So instead of using a single transaction this change
uses separate transaction for scheduling each message email.