Bad request is returned if createOrganization does not create a new organization entity.
There fixes a bug that even if result equaled to False, its extra part would be passed to createConnection function. It was not right, because the function expected Organization entity, whereas result.extra pointed to unicode. The following error would be raised:

386, in createConnectionTxn
    can_create = connection_logic.canCreateConnection(profile, organization.key)
AttributeError: 'unicode' object has no attribute 'key'

It fixes the test case which was introduced in the previous revision.