Organization entity is retrived within a transaction, not passed to the transaction.
It is an improvement, because transactions should modify fresh data. Moreorver, get() in NDB world may obtain stale data from in memory cache or memcache. What is interesting, the transaction failed before this change was applied. It actually printed the same error as the one which was described in the log message to revision 6d31e5d78b73.