executeMapReduceJobs function added to test utils.
MapReduce in AppEngine is implemented on the top of Task Queue API. Normally we should be able to execute these tasks directly, using the pattern like in executeTasks function. Unfortunately, the tasks for MapReduce have some sort of different format and I was not able to do that.

Fortunately, test_support provides execute_until_empty function which does the job.

BTW. We use 'default' queue for map reduce tasks. Therefore, this function will actually execute (probably) all tasks in that queue, not only map reduce ones. It should not be a practical problem.