The form, which wraps the button to finalize decisions, has 'form-finalize-decision' class.
Also, CSS is added for this class in conjuction with "form" id. This is why I loathe our forms and CSS rules which are defined for it. Basically, the CSS exects that each form has id "form". What if there are more forms on the page? They both need to share the same ID. What if we want to define a custom form that shares *some* properties of generic forms? We need to do what I have done in this revision: create a new class, apply it to the form element along with id="form" and define CSS rule for #form.class_name. It is necessary because ID bind much stronger than classes. In other words, when CSS is applied and the interpreter sees one rule for #form and another one for .class_name, it does not find the latter one important at all :-(

Margin 0px is applied in order to adjust the button to the list below. Without it, Chrome applies 2px by default and the button is shifted a bit to the right.

The main reason for this change is issue 2008: the button is too close to the list. Adding margin-botton resolves the problem.