createConnectionTxn function is changed so that welcome email may be requested.
It is not used right now, but it may be useful at some point, when anonymous connections are supported again. At this moment, this function is called in one of three cases:
1) the user starts a connection
2) the organization starts a connection
3) the organization profile is created a connection is started with the main admin and the backup admin.

In case of (1) and (2), no role is actually granted, beuacse the second side, either organization or the user must accept the connection. So, the email is not needed.
In case of (3) Org Admin role is assigned to the main admin and the backup admin, but the organiztion is not accepted yet (they are only applying), so no email is needed.