Fixes to getAcceptedOrganizationsWithLogoURLs.
The break inside the for loop was always meant to be
a break out of the enclosing while loop. While Python
doesn't have labeled break statements, it also doesn't
have magic break statements either. I don't know what
I was thinking. A local field "found_org_found_again"
is used to simulate the effect of a labeled break.

It wasn't right that the length of "orgs", which could
change during the middle of the while loop, was used
at the end of the loop as though it hadn't changed. A
local field "requested" is added to be consistent
about exactly how many orgs were requested in the
query at the beginning of the loop.

It wasn't right that the cursor is reassigned at the
top of the loop and then used in comparisons at the
bottom of the loop as though it had kept its original
value. A local field "next_cursor" is introduced to
allow "cursor" to retain the same value at all sites
where it is read in a single loop iteration.