Order for the query to fetch connections for users who admin more than one organization is explicitly set to _key.
Otherwise, fetch_page fails with the following error:
BadArgumentError: _MultiQuery with cursors requires __key__ order

We could add something like
if self._query.is_multi_query:
  self._query = self._query.order(ndb.Model._key)

but Query class does not provide a public function for that. There is
  def _needs_multi_query(self):
    filters = self.filters
    return filters is not None and isinstance(filters, DisjunctionNode)
but it is private.

Another option is to implement it ourselves or report it to AppEngine.