Inject linker and url_names instances from the mapreduce to organization application decision logic.
This should have ideally been injected from the code that initiates the
mapreduce. However, I chose against that because the form of communicating
between the code that initiates the mapreduce and the mapreduce code
itself is only through serialized data. I am not sure if mapreduce Python
API does the serialization-deserialization, even if it does it does not
make sense to be serializing these instance objects and deserializing it
for every instance of the mapper run.