rejected_for field is added to Profile model.
It lists organizations which the specified profile administers but which are not accepted into the program. Such users may still have some privileges. For example, they may want to see their organization application response.

One of the options was not to distinguish between accepted admins and rejected admins. In majority of cases there is no need to do that. On the other hand, we want to filter actual organization admins who actually participate in the program. Program administrators need to know who participates in the program and who does not.